fix: atom,layer 폴더링, index 파일들 수정, 경로수정
