+++ src/components/index.jsx
@@ -1,13 +1,30 @@
-//atoms
-export { default as Header } from "./common/Header";
+// atoms
 export { default as Swiper } from "./atoms/Swiper";
-export { default as Footer } from './common/Footer';
 export { default as DogDetailInfo } from "./atoms/DogDetailInfo";
 export { default as DogDetail } from "./atoms/DogDetail";
 export { default as FindDogCard } from "./atoms/FindDogCard";
+export { default as Searchbar } from "./atoms/Searchbar";
 
-//molcules
-export { default as DogCardContainer } from "./molecules/DogCardContainer";
-export { default as PaginationNav } from "./molecules/PaginationNav";
-export { default as CarouselDogContainer } from "./layer/mainPage/CarouselDogContainer";
-export { default as CarouselReviewContainer } from "./layer/mainPage/CarouselReviewContainer";
+// layer
+export { default as Header } from "./layer/Header";
+export { default as Footer } from './layer/Footer';
+
+// layer/mainPage
+export { default as AboutTakeUs } from './layer/mainPage/AboutTakeUs';
+export { default as BeforeVolunteer } from './layer/mainPage/BeforeVolunteer';
+export { default as CarouselDogContainer } from './layer/mainPage/CarouselDogContainer';
+export { default as CarouselReviewContainer } from './layer/mainPage/CarouselReviewContainer';
+
+// layer/about
+
+// layer/dogEnroll
+
+// layer/dogSearch
+export { default as DogCardContainer } from "./layer/dogSearch/DogCardContainer";
+export { default as PaginationNav } from "./layer/dogSearch/PaginationNav";
+
+// layer/info
+
+// layer/review
+
+// layer/login
