+++ src/components/layer/about/about.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const about = () => {
+  return (
+    <div>
+      aboutLayer
+    </div>
+  );
+};
+
+export default about;
