+++ src/components/layer/dogEnroll/dogEnrollLayer.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const dogEnrollLayer = () => {
+  return (
+    <div>
+      dogEnrollLayer
+    </div>
+  );
+};
+
+export default dogEnrollLayer;
