+++ src/components/layer/info/infoLayer.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const infoLayer = () => {
+  return (
+    <div>
+      infoLayer
+    </div>
+  );
+};
+
+export default infoLayer;
