+++ src/components/layer/login/loginLayer.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const loginLayer = () => {
+  return (
+    <div>
+      loginLayer
+    </div>
+  );
+};
+
+export default loginLayer;
