+++ src/pages/MainPage.jsx
@@ -1,10 +1,6 @@
 import React from "react";
 import styled from 'styled-components';
-import AboutTakeUs from 'components/layer/mainPage/AboutTakeUs';
-import BeforeVolunteer from '../components/layer/mainPage/BeforeVolunteer';
-import CarouselDogContainer from "../components/layer/mainPage/CarouselDogContainer";
-import CarouselReviewContainer from "../components/layer/mainPage/CarouselReviewContainer";
-import Searchbar from '../components/common/Searchbar';
+import { Searchbar, AboutTakeUs, BeforeVolunteer, CarouselDogContainer, CarouselReviewContainer } from '../components';
 
 const Styled = {
   Wrapper: styled.div`
