+++ src/pages/about/aboutPage.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ContactPage = () => {
+  return <div>about</div>;
+};
+
+export default ContactPage;
