--- src/pages/contact/ContactPage.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ContactPage = () => {
-  return <div>contact</div>;
-};
-
-export default ContactPage;
