--- src/pages/dog/DogDetailPage.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import { DogDetail } from '../../components';
-
-const DogDetailPage = () => {
-  return (
-    <section>
-      <DogDetail />
-    </section>
-  );
-};
-
-export default DogDetailPage;
