--- src/pages/dog/DogEnrollPage.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const DogEnrollPage = () => {
-  return <div>대상견 등록</div>;
-};
-
-export default DogEnrollPage;
