+++ src/pages/dogEnroll/DogEnrollPage.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const DogEnrollPage = () => {
+  return <div>대상견 등록</div>;
+};
+
+export default DogEnrollPage;
