+++ src/pages/dogSearch/DogDetailPage.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { DogDetail } from '../../components';
+
+const DogDetailPage = () => {
+  return (
+    <section>
+      <DogDetail />
+    </section>
+  );
+};
+
+export default DogDetailPage;
