+++ src/pages/index.jsx
@@ -1,7 +1,7 @@
-export { default as DogPage } from "./dog/DogPage";
-export { default as DogDetailPage } from "./dog/DogDetailPage";
-export { default as DogEnrollPage } from "./dog/DogEnrollPage";
-export { default as ContactPage } from "./contact/ContactPage";
+export { default as DogPage } from "./dogSearch/DogPage";
+export { default as DogDetailPage } from "./dogSearch/DogDetailPage";
+export { default as DogEnrollPage } from "./dogEnroll/DogEnrollPage";
+export { default as AboutPage } from "./about/aboutPage";
 export { default as InfoPage } from "./info/InfoPage";
 export { default as LoginPage } from "./login/LoginPage";
 export { default as ReviewPage } from "./review/ReviewPage";
